+++ lumicks/pylake/detail/image.py
@@ -2,6 +2,7 @@ import enum
 import math
 import numpy as np
 import warnings
+from deprecated import deprecated
 
 
 class InfowaveCode(enum.IntEnum):
@@ -173,54 +174,6 @@ def reconstruct_image(data, infowave, shape, reduce=np.sum):
     resized_data = data[valid_idx][: round_down(subset.size, pixel_size)]
     pixels = reduce(resized_data.reshape(-1, pixel_size), axis=1)
     return reshape_reconstructed_image(pixels, shape)
-
-
-def save_tiff(
-    image, filename, dtype, clip=False, pixel_sizes_um=(1e-3, 1e-3), pixel_time_seconds=1.0
-):
-    """Save an RGB `image` to TIFF
-
-    This is a thin wrapper around `tifffile` with additional safety checks
-
-    Parameters
-    ----------
-    image : np.array
-        Image data. An RGB image is expected to have `shape == (w, h, 3)`.
-    filename : str
-        The name of the TIFF file where the image will be saved.
-    dtype : np.dtype
-        The data type of a single color channel in the resulting image.
-    clip : bool
-        If enabled, the photon count data will be clipped to fit into the desired `dtype`.
-        This option is disabled by default: an error will be raise if the data does not fit.
-    pixel_sizes_um : array_like
-        Pixel sizes in [um]
-    pixel_time_seconds : float
-        Pixel acquisition time [s]
-    """
-    import tifffile
-
-    info = np.finfo(dtype) if np.dtype(dtype).kind == "f" else np.iinfo(dtype)
-    if not clip and (np.min(image) < info.min or np.max(image) > info.max):
-        raise RuntimeError(
-            f"Can't safely export image with `dtype={dtype.__name__}` channels."
-            f" Switch to a larger `dtype` in order to safely store everything"
-            f" or pass `force=True` to clip the data."
-        )
-
-    pixel_size_x, pixel_size_y = (
-        pixel_sizes_um[0],
-        pixel_sizes_um[1] if len(pixel_sizes_um) == 2 else pixel_sizes_um[0],
-    )
-
-    tiff_kwargs = {
-        "metadata": {"PixelTime": pixel_time_seconds, "PixelTimeUnit": "s"},
-        "photometric": "rgb",
-    }
-    if pixel_size_x:
-        tiff_kwargs["resolution"] = (1e4 / pixel_size_x, 1e4 / pixel_size_y, "CENTIMETER")
-
-    tifffile.imwrite(filename, image.astype(dtype), **tiff_kwargs)
 
 
 def histogram_rows(image, pixels_per_bin, pixel_width):